# Add locks to param config