--- a/shared/params/config.go
+++ b/shared/params/config.go
@@ -2,6 +2,7 @@
 package params
 
 import (
+	"sync"
 	"time"
 
 	"github.com/mohae/deepcopy"
@@ -119,9 +120,12 @@ type BeaconChainConfig struct {
 }
 
 var beaconConfig = MainnetConfig()
+var lock sync.RWMutex
 
 // BeaconConfig retrieves beacon chain config.
 func BeaconConfig() *BeaconChainConfig {
+	lock.RLock()
+	defer lock.RUnlock()
 	return beaconConfig
 }
 
@@ -130,11 +134,15 @@ func BeaconConfig() *BeaconChainConfig {
 // OverrideBeaconConfig(c). Any subsequent calls to params.BeaconConfig() will
 // return this new configuration.
 func OverrideBeaconConfig(c *BeaconChainConfig) {
+	lock.Lock()
+	defer lock.Unlock()
 	beaconConfig = c
 }
 
 // Copy returns a copy of the config object.
 func (c *BeaconChainConfig) Copy() *BeaconChainConfig {
+	lock.Lock()
+	defer lock.Unlock()
 	config, ok := deepcopy.Copy(*c).(BeaconChainConfig)
 	if !ok {
 		config = *beaconConfig