--- a/shared/params/mainnet_config.go
+++ b/shared/params/mainnet_config.go
@@ -13,6 +13,8 @@ func MainnetConfig() *BeaconChainConfig {
 
 // UseMainnetConfig for beacon chain services.
 func UseMainnetConfig() {
+	lock.Lock()
+	defer lock.Unlock()
 	beaconConfig = MainnetConfig()
 }