--- a/shared/params/testnet_e2e_config.go
+++ b/shared/params/testnet_e2e_config.go
@@ -19,7 +19,7 @@ func UseE2EConfig() {
 // E2ETestConfig retrieves the configurations made specifically for E2E testing.
 // Warning: This config is only for testing, it is not meant for use outside of E2E.
 func E2ETestConfig() *BeaconChainConfig {
-	lock.RUnlock()
+	lock.RLock()
 	defer lock.RUnlock()
 	return e2ETestConfig()
 }