--- a/shared/params/testnet_pyrmont_config.go
+++ b/shared/params/testnet_pyrmont_config.go
@@ -18,6 +18,8 @@ func UsePyrmontNetworkConfig() {
 // UsePyrmontConfig sets the main beacon chain
 // config for Pyrmont.
 func UsePyrmontConfig() {
+	lock.Lock()
+	defer lock.Unlock()
 	beaconConfig = PyrmontConfig()
 }