<!-- Thanks for sending a PR! Before submitting:

1. If this is your first PR, check out our contribution guide here https://docs.prylabs.network/docs/contribute/contribution-guidelines
   You will then need to sign our Contributor License Agreement (CLA), which will show up as a comment from a bot in this pull request after you open it. We cannot review code without a signed CLA.
2. Please file an associated tracking issue if this pull request is non-trivial and requires context for our team to understand. All features and most bug fixes should have
   an associated issue with a design discussed and decided upon. Small bug
   fixes and documentation improvements don't need issues.
3. New features and bug fixes must have tests. Documentation may need to
   be updated. If you're unsure what to update, send the PR, and we'll discuss
   in review.
4. Note that PRs updating dependencies and new Go versions are not accepted.
   Please file an issue instead.
-->

**What type of PR is this?**

> Feature

**What does this PR do? Why is it needed?**

We should add lock around param config usages. Some tests that I wrote in #8094 failed race because of it. 

Race log:
```

WARNING: DATA RACE
--
  | Write at 0x000002eb6aa0 by goroutine 44:
  | github.com/prysmaticlabs/prysm/shared/params.OverrideBeaconConfig()
  | shared/params/config.go:133 +0x145
  | github.com/prysmaticlabs/prysm/beacon-chain/blockchain.TestService_ReceiveBlock()
  | beacon-chain/blockchain/receive_block_test.go:35 +0x135
  | testing.tRunner()
  | GOROOT/src/testing/testing.go:1123 +0x202
  |  
  | Previous read at 0x000002eb6aa0 by goroutine 32:
  | github.com/prysmaticlabs/prysm/beacon-chain/blockchain.(*Service).HeadRoot()
  | beacon-chain/blockchain/chain_info.go:125 +0x3f4
  | github.com/prysmaticlabs/prysm/beacon-chain/blockchain.(*Service).saveHead()
  | beacon-chain/blockchain/head.go:86 +0x117
  | github.com/prysmaticlabs/prysm/beacon-chain/blockchain.TestHeadBlock_DataRace.func1()
  | beacon-chain/blockchain/chain_info_norace_test.go:46 +0x88
```

**Which issues(s) does this PR fix?**

Discovered in #8094

**Other notes for review**