@terencechain 2020-12-15
> I dont know if this is such a good idea. Adding in a lock for our config, basically puts it in the hot path for everything. Nearly every method in our core package/services will be accessing the same config at the same time. They will all also have to acquire/release the lock here, bringing additional overhead.  

For the read locks, concurrent reads can happen with negligible performance impact.

This lock is proposed because there's a race failure with when we overwrite config in tests, in the future, someone could also try this at run time, I do prefer safety.