@nisdas 2020-12-15
> For the read locks, concurrent reads can happen with negligible performance impact.

Are there any benchmarks to prove this ?

> This lock is proposed because there's a race failure with when we overwrite config in tests, in the future, someone could also try this at run time, I do prefer safety.

Wouldn't that be a test setup issue then ? From the failure it looks like its because you have multiple tests running at the same time, shouldn't tests not be run in parallel then.