@terencechain 2020-12-15
> Are there any benchmarks to prove this ?

Can certainly write a benchmark to further prove this.

> Wouldn't that be a test setup issue then ? From the failure it looks like its because you have multiple tests running at the same time, shouldn't tests not be run in parallel then.

It's a test setup issue for now, but the worry here is it lingers long enough to become a run time / production issue. Imagine a simple code change that I made and was able to trigger this race failure with one test. This params package does not look safe to me without proper locking