@terencechain 2020-12-15
Here's the benchmark.
```go
func Benchmark_GetConfig(b *testing.B) {
	b.RunParallel(func(pb *testing.PB) {
		for pb.Next() {
			if params.BeaconConfig().SlotsPerEpoch != 32 {
				b.Fatal("Incorrect SlotsPerEpoch")
			}
		}
	})
}
```

Without read lock:
pkg: github.com/prysmaticlabs/prysm/shared/params
Benchmark_GetConfig
Benchmark_GetConfig-12    	1000000000	         0.165 ns/op
PASS

With read lock:
pkg: github.com/prysmaticlabs/prysm/shared/params
Benchmark_GetConfig
Benchmark_GetConfig-12    	30282727	        34.5 ns/op
PASS

About 30ns slower. Question now is whether 30ns is an acceptable trade off for safety