@nisdas 2020-12-15
> It's a test setup issue for now, but the worry here is it lingers long enough to become a run time / production issue. Imagine a simple code change that I made and was able to trigger this race failure with one test, how likely it could happen again in the future. This params package does not look safe to me without proper locking

So the fundamental issue we must address is that do we expect to write to/initialize the config while the node is running normally (after node initialization) ? If there is the possibility that it can happen while the node is up and running normally and we do try to modify the config, we will need a lock here. But in the event that its possible for something like that to happen, we probably need to re-check our assumptions. We assume that the config is a static object for the most part, if it can be continuously modified in runtime it means a lot of things that we assume when running our services don't hold. 

> Without read lock:
>
>pkg: github.com/prysmaticlabs/prysm/shared/params
Benchmark_GetConfig
Benchmark_GetConfig-12    	1000000000	         0.165 ns/op
PASS
With read lock:

> pkg: github.com/prysmaticlabs/prysm/shared/params
Benchmark_GetConfig
Benchmark_GetConfig-12    	30282727	        34.5 ns/op
PASS
About 30ns slower. Question now is whether 30ns is an acceptable trade off for safety

Thats a pretty big difference, even though this is only 1 value it all eventually adds up . Probably a better benchmark is to benchmark state transitions with/without the config. Since that is our most important code path to be concerned of.