@terencechain 2020-12-15
> So the fundamental issue we must address is that do we expect to write to/initialize the config while the node is running normally (after node initialization)

It's not so much on expection, it's more on allowence. We can say that we don't expect one to modify config during run time but anyone can make a mistake doing it and we can miss reviewing it. The best solution us don't allow config overwritr and treat config as static file. The next best solution is use a lock. Do nothing is not good long term imo