@terencechain 2020-12-15
As promised, here's some benchmark numbers for full state transtion.

Without locking:
BenchmarkExecuteStateTransition_FullProcess-12          18  66811454 ns/op

With locking:
BenchmarkExecuteStateTransition_FullProcess-12    	      16	  67013178 ns/op

It's about 0.3% worse than before

Source:
```go
func BenchmarkExecuteStateTransition_FullProcess(b *testing.B) {
	beaconState, privKeys := testutil.DeterministicGenesisState(b, 100)

	eth1Data := &ethpb.Eth1Data{
		DepositCount: 100,
		DepositRoot:  bytesutil.PadTo([]byte{2}, 32),
		BlockHash:    make([]byte, 32),
	}
	require.NoError(b, beaconState.SetSlot(params.BeaconConfig().SlotsPerEpoch-1))
	e := beaconState.Eth1Data()
	e.DepositCount = 100
	require.NoError(b, beaconState.SetEth1Data(e))
	bh := beaconState.LatestBlockHeader()
	bh.Slot = beaconState.Slot()
	require.NoError(b, beaconState.SetLatestBlockHeader(bh))
	require.NoError(b, beaconState.SetEth1DataVotes([]*ethpb.Eth1Data{eth1Data}))

	require.NoError(b, beaconState.SetSlot(beaconState.Slot()+1))
	epoch := helpers.CurrentEpoch(beaconState)
	randaoReveal, err := testutil.RandaoReveal(beaconState, epoch, privKeys)
	require.NoError(b, err)
	require.NoError(b, beaconState.SetSlot(beaconState.Slot()-1))

	nextSlotState, err := state.ProcessSlots(context.Background(), beaconState.Copy(), beaconState.Slot()+1)
	require.NoError(b, err)
	parentRoot, err := nextSlotState.LatestBlockHeader().HashTreeRoot()
	require.NoError(b, err)
	proposerIdx, err := helpers.BeaconProposerIndex(nextSlotState)
	require.NoError(b, err)
	block := testutil.NewBeaconBlock()
	block.Block.ProposerIndex = proposerIdx
	block.Block.Slot = beaconState.Slot() + 1
	block.Block.ParentRoot = parentRoot[:]
	block.Block.Body.RandaoReveal = randaoReveal
	block.Block.Body.Eth1Data = eth1Data

	stateRoot, err := state.CalculateStateRoot(context.Background(), beaconState, block)
	require.NoError(b, err)

	block.Block.StateRoot = stateRoot[:]

	sig, err := testutil.BlockSignature(beaconState, block.Block, privKeys)
	require.NoError(b, err)
	block.Signature = sig.Marshal()

	b.ResetTimer()
	for i := 0; i < b.N; i++ {
		copied := beaconState.Copy()
		_, err = state.ExecuteStateTransition(context.Background(), copied, block)
		require.NoError(b, err)
	}
}
```