@dshulyak 2020-12-20
if 30ns of overhead is your primary concern and updates are rare you can use this pattern with atomics:

```go
import (
	"sync/atomic"
	"unsafe"
)

var cfg unsafe.Pointer = unsafe.Pointer(&Config{})

type Config struct {
	value string
}

func getConfig() *Config {
	return (*Config)(atomic.LoadPointer(&cfg))
}

func updateConfig(f func(c *Config)) {
	for {
		c1 := getConfig()
		c2 := &Config{value: c1.value} // copy
		f(c2)
		if atomic.CompareAndSwapPointer(&cfg, unsafe.Pointer(c1), unsafe.Pointer(c2)) {
			return
		}
	}
}
```

getConfig has no overhead at all on x86 (will probably have insignificant on systems with weaker memory model)